+++ code/Your_Story/ysProject/src/mainPackage/ChatConnection.java
@@ -25,7 +25,7 @@ public class ChatConnection {
         r = DBInterface.getConnection().selectStuff("SELECT body FROM " + CHAT_DATA + " WHERE lobby = " + lobby + " AND id > " + lastMessage);
         return DBInterface.resultSetToStringArray(r);
     }
-    public static int sendMessage(String message, int lobby){
+    public static long sendMessage(String message, int lobby){
         message = DBInterface.escapeString(message);
         DBInterface.getConnection().executeStuff("INSERT INTO " + CHAT_DATA + "(body, lobby) VALUES('" + message + "', " + lobby + ")");
         long[] temp = DBInterface.selectIntArray(CHAT_DATA, "id", "lobby", lobby);
